+++ app/controllers/developers_controller.rb
@@ -66,6 +66,7 @@ class DevelopersController < ApplicationController
       :cover_image,
       :search_status,
       :search_query,
+      :profile_reminder_notifications,
       location_attributes: [:city, :state, :country],
       role_type_attributes: RoleType::TYPES,
       role_level_attributes: RoleLevel::TYPES