+++ db/migrate/20220523115740_add_send_stale_notification_to_developers.rb
@@ -1,0 +1,5 @@
+class AddSendStaleNotificationToDevelopers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :developers, :profile_reminder_notifications, :boolean, default: true
+  end
+end