+++ test/mailers/previews/developer_mailer_preview.rb
@@ -4,6 +4,11 @@ class DeveloperMailerPreview < ActionMailer::Preview
     DeveloperMailer.with(record: notification, recipient: notification.recipient).invisiblize
   end
 
+  def stale
+    notification = Notification.where(type: StaleDeveloperNotification.to_s).first
+    DeveloperMailer.with(record: notification, recipient: notification.recipient).stale
+  end
+
   def welcome
     DeveloperMailer.with(developer: Developer.first).welcome_email
   end